+++ src/resources/views/filters/select2_ajax.blade.php
@@ -19,14 +19,11 @@
 
 @push('crud_list_styles')
     <!-- include select2 css-->
-    <link href="{{ asset('vendor/adminlte/bower_components/select2/dist/css/select2.min.css') }}" rel="stylesheet" type="text/css" />
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/select2-bootstrap-theme/0.1.0-beta.10/select2-bootstrap.min.css" rel="stylesheet" type="text/css" />
+    <link href="{{ asset('vendor/backpack/select2/select2.css') }}" rel="stylesheet" type="text/css" />
+    <link href="{{ asset('vendor/backpack/select2/select2-bootstrap-dick.css') }}" rel="stylesheet" type="text/css" />
     <style>
 	  .form-inline .select2-container {
 	    display: inline-block;
-	  }
-	  li[filter-type="{{ $filter->type }}"] .select2-container {
-	  	display: block;
 	  }
 	  .select2-drop-active {
 	  	border:none;
@@ -47,7 +44,7 @@
 
 @push('crud_list_scripts')
 	<!-- include select2 js-->
-    <script src="{{ asset('vendor/adminlte/bower_components/select2/dist/js/select2.min.js') }}"></script>
+    <script src="{{ asset('vendor/backpack/select2/select2.js') }}"></script>
     <script>
         jQuery(document).ready(function($) {
             // trigger select2 for each untriggered select2 box